chore: move sync-cac-catalog to sync-cac-content
