+++ trestlebot/cli/root.py
@@ -10,8 +10,10 @@ from trestlebot.cli.commands.autosync import autosync_cmd
 from trestlebot.cli.commands.create import create_cmd
 from trestlebot.cli.commands.init import init_cmd
 from trestlebot.cli.commands.rules_transform import rules_transform_cmd
-from trestlebot.cli.commands.sync_cac_catalog import sync_cac_catalog_cmd
-from trestlebot.cli.commands.sync_cac_content import sync_cac_content_cmd
+from trestlebot.cli.commands.sync_cac_content import (
+    sync_cac_catalog_cmd,
+    sync_cac_content_cmd,
+)
 from trestlebot.cli.commands.sync_upstreams import sync_upstreams_cmd
